feat(sandbox.ts): add sandbox.ts file to generate a URL with query parameters for a code snippet
The sandbox.ts file is added to the project. It imports the "qs" library for query string manipulation. It defines an options object with a "code" property containing a JavaScript function. It then generates a URL using the "qs.stringify" method to serialize the options object as query parameters. Finally, it logs the generated URL to the console. This file is added to provide a convenient way to generate URLs for code snippets in the project.